@@ -3,31 +3,54 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "axios": "^0.18.0",
-    "lodash": "^4.17.11",
-    "node-sass": "^4.11.0",
-    "npm": "^6.6.0",
-    "react": "^16.7.0",
+    "@types/jest": "^24.0.15",
+    "@types/node": "^12.6.0",
+    "@types/react": "^16.8.22",
+    "@types/react-dom": "^16.8.4",
+    "@types/react-redux": "^7.1.1",
+    "@types/react-router-dom": "^4.3.4",
+    "@types/redux-form": "^8.1.5",
+    "art": "^0.10.3",
+    "axios": "^0.19.0",
+    "cropperjs": "^1.5.5",
+    "date-fns": "^2.0.0",
+    "final-form": "^4.18.5",
+    "lodash": "^4.17.14",
+    "node-sass": "^4.12.0",
+    "prop-types": "^15.7.2",
+    "query-string": "^6.8.2",
+    "react": "^16.9.0",
     "react-bulma-components": "^3.0.1-2",
-    "react-dom": "^16.7.0",
+    "react-country-region-selector": "^1.4.3",
+    "react-dom": "^16.9.0",
+    "react-final-form": "^6.3.0",
     "react-identicons": "^1.1.7",
     "react-ionicons": "^3.0.1",
-    "react-phone-input-2": "^2.7.1",
-    "react-qr-reader": "^2.1.2",
-    "react-redux": "^6.0.0",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "^2.1.3",
-    "react-swipe": "^6.0.3",
-    "redux": "^4.0.1",
-    "redux-form": "^8.0.4",
+    "react-phone-input-2": "^2.9.5",
+    "react-qr-code": "^0.1.3",
+    "react-qr-reader": "^2.2.1",
+    "react-redux": "^7.1.0",
+    "react-router-dom": "^5.0.1",
+    "react-scripts": "^3.1.0",
+    "react-spring": "^8.0.27",
+    "react-swipe": "^6.0.4",
+    "react-transition-group": "^4.2.2",
+    "react-use-gesture": "^5.2.4",
+    "react-virtualized-auto-sizer": "^1.0.2",
+    "react-window": "^1.8.5",
+    "redux": "^4.0.4",
+    "redux-form": "^8.2.6",
     "redux-thunk": "^2.3.0",
-    "swipe-js-iso": "^2.1.4"
+    "swipe-js-iso": "^2.1.5",
+    "typescript": "^3.5.3"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "styleguide": "styleguidist server",
+    "styleguide:build": "styleguidist build"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -37,5 +60,8 @@
     "not dead",
     "not ie <= 11",
     "not op_mini all"
-  ]
+  ],
+  "devDependencies": {
+    "react-styleguidist": "^9.1.2"
+  }
 }
